+++ tests/utils/pixInfraction.py
@@ -8,7 +8,9 @@ from starkinfra import PixInfraction, pixinfraction
 example_pix_infraction = PixInfraction(
     reference_id=get_end_to_end_id()[0],
     type="reversal",
-    method="scam"
+    method="scam",
+    operator_email="fraud@company.com",
+    operator_phone="+5511999999999",
 )
 
 
@@ -28,7 +30,7 @@ def getPixInfractionToPatch():
     while len(infraction_reports) < 1:
         reports, cursor = pixinfraction.page(status="delivered", limit=1, cursor=cursor)
         for report in reports:
-            if report.flow == "out":
+            if report.flow == "in":
                 infraction_reports.append(report)
         if cursor is None:
             break